feat : 게시판 현 주인만 삭제가 가능하도록 조건 추가

+++ community/src/main/java/efub/assignment/community/board/service/BoardService.java
@@ -4,6 +4,8 @@ import efub.assignment.community.board.domain.Board;
 import efub.assignment.community.board.dto.BoardRequestDto;
 import efub.assignment.community.board.dto.BoardUpdateRequestDto;
 import efub.assignment.community.board.repository.BoardRepository;
+import efub.assignment.community.exception.CustomDeleteException;
+import efub.assignment.community.exception.ErrorCode;
 import efub.assignment.community.member.domain.Member;
 import efub.assignment.community.member.service.MemberService;
 import jakarta.persistence.EntityNotFoundException;
@@ -11,6 +13,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import static efub.assignment.community.exception.ErrorCode.DELETE_PERMISSION_REJECTED_USER;
+import static efub.assignment.community.exception.ErrorCode.UPDATE_PERMISSION_REJECTED_USER;
 
 @Service
 @Transactional
@@ -36,14 +40,18 @@ public class BoardService {
     public Long update(Long boardId, BoardUpdateRequestDto requestDto){
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(()-> new EntityNotFoundException("해당 id를 가진 board를 찾을 수 없습니다."));
-        Member member = memberService.findMemberById(Long.parseLong(requestDto.getMemberId()));
+        Long memberId = Long.parseLong(requestDto.getMemberId());
+        Member member = memberService.findMemberById(memberId);
         board.updateBoard(member);
         return board.getBoardId();
     }
 
-    public void deleteBoard(Long id){
+    public void deleteBoard(Long id, Long memberId){
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("해당 id를 가진 Board를 찾을 수 없습니다. id="+id));
+        if(!memberId.equals(board.getMember().getMemberId())){ // 게시판 주인이 아닌 경우 삭제 권한 없음 처리
+            throw new CustomDeleteException(DELETE_PERMISSION_REJECTED_USER);
+        }
         boardRepository.delete(board);
     }
 }
